fix: replace factory with separate methods
