+++ app/components/inspector/utils.ts
@@ -1,17 +1,11 @@
-import * as spl from '@solana/spl-token';
 import {
     AccountMeta,
     MessageAddressTableLookup,
     MessageCompiledInstruction,
-    ParsedInstruction,
-    ParsedMessage,
-    ParsedMessageAccount,
-    ParsedTransaction,
     PublicKey,
     TransactionInstruction,
     VersionedMessage,
 } from '@solana/web3.js';
-import { CREATE_ASSOCIATED_TOKEN_DISCRIMINATOR, CREATE_ASSOCIATED_TOKEN_IDEMPOTENT_DISCRIMINATOR, RECOVER_NESTED_ASSOCIATED_TOKEN_DISCRIMINATOR } from '@solana-program/token';
 
 type LookupsForAccountKeyIndex = { lookupTableIndex: number, lookupTableKey: PublicKey }
 
@@ -107,100 +101,4 @@ export function intoTransactionInstructionFromVersionedMessage(
     });
 
     return transactionInstruction;
-}
-
-/**
- * entity that performs conversion from TransactionInstruction that is created from VersionedMessage into ParsedInstruction.
- *
- *  That is needed for backward compatibility with existing InstructionCards that expect ParsedInstruction for rendering.
- *
- *  This is temporary solution to reuse existing cards at the Inspector. Might pivot to a better solution in future.
- */
-export function ParsedInstructionFactory(){
-    function discriminatorToBuffer(discrimnator: number): Buffer{
-        return Buffer.from(Uint8Array.from([discrimnator]));
-    }
-
-    function intoProgramName(programId: PublicKey): string | undefined {
-        if (programId.equals(spl.ASSOCIATED_TOKEN_PROGRAM_ID)) {
-            return 'spl-associated-token-account';
-        }
-    }
-
-    function intoParsedData(instruction: TransactionInstruction, parsed: any): any{
-        const { programId, data } = instruction;
-        const info = parsed ?? {};
-
-        if (programId.equals(spl.ASSOCIATED_TOKEN_PROGRAM_ID)) {
-            let type;
-            if (data.equals(Buffer.alloc(CREATE_ASSOCIATED_TOKEN_DISCRIMINATOR))) type = 'create';
-            else if (data.equals(discriminatorToBuffer(CREATE_ASSOCIATED_TOKEN_DISCRIMINATOR))) type = 'create';
-            else if (data.equals(discriminatorToBuffer(CREATE_ASSOCIATED_TOKEN_IDEMPOTENT_DISCRIMINATOR))) type = 'createIdempotent';
-            else if (data.equals(discriminatorToBuffer(RECOVER_NESTED_ASSOCIATED_TOKEN_DISCRIMINATOR))) type ='recoverNested';
-            else type = '';
-
-            return {
-                info,
-                type
-            };
-        }
-
-        return {
-            info,
-            type: '' // empty string represents that the program was unknown
-        };
-    }
-
-    function getInstructionData(instruction: TransactionInstruction, data: any){
-        const program = intoProgramName(instruction.programId);
-        const parsed = intoParsedData(instruction, data);
-
-        return { parsed, program };
-    }
-
-    function convertAccountKeysToParsedMessageAccounts(keys: AccountMeta[]): ParsedMessageAccount[]{
-        const accountKeys= keys.map((key): ParsedMessageAccount => {
-            return {
-                pubkey: key.pubkey,
-                signer: key.isSigner,
-                source: 'lookupTable',
-                writable: key.isWritable
-            };
-        });
-
-        return accountKeys;
-    }
-
-    return {
-        /**
-         * parsed - allow to pass any data as we can not recover parsed data
-         */
-        intoParsedInstruction(transactionInstruction: TransactionInstruction, data: any = {}): ParsedInstruction {
-            const { programId } = transactionInstruction;
-            const { program, parsed } = getInstructionData(transactionInstruction, data);
-
-            return {
-                parsed,
-                program: program ?? '',
-                programId
-            };
-        },
-        intoParsedTransaction(transactionInstruction: TransactionInstruction, versionedMessage: VersionedMessage): ParsedTransaction {
-            const { keys } = transactionInstruction;
-            const { addressTableLookups, recentBlockhash } = versionedMessage;
-
-            const parsedMessage: ParsedMessage = {
-                accountKeys: convertAccountKeysToParsedMessageAccounts(keys),
-                addressTableLookups,
-                // at this moment we do not parse instructions as they are not required to represent the transaction. that's why array is empty
-                instructions: [],
-                recentBlockhash
-            };
-
-            return {
-                message: parsedMessage,
-                signatures: [],
-            };
-        }
-    };
 }
